Change OO design, add UI using curses, create input functions
Student and Course class are now only responsible for storing
data: id, name, etc and setters and getters

Add two classes Mark and Department. Mark for storing
mark of a specific student in a specific course. Also, when create a new
Mark instance, it is automatically added to student and course
accordingly. Department have student list, course list, mark list and
is responsible for some  utility functions as calculate GPA which used
to be in Student and Course class

Add UI design using curses module. Initially, it can show list of
students, list of courses and list of marks depending on user choice.

Input functionalities used to be in the main functions. Now, they are
put in specific fuctions ,e.g.: input_students